# docs/openapi.yaml
openapi: 3.1.0
info:
  title: smalltalk API
  version: 0.2.0
  summary: Room chat joined by invite link or QR. No app install, no sign-up.
  description: |
    REST API and WebSocket contract of the smalltalk server (`backend/`).

    **Conventions**
    - JSON fields are `snake_case`; times are RFC 3339 in UTC, e.g. `2026-09-26T06:00:00Z`.
    - Every error uses the same envelope: `{"error": {"code": "invite_expired", "message": "..."}}`.
      Clients should branch on `code`, never on `message`.
    - Authentication is a short-lived JWT (12 h) bound to **one room and one member**. It is returned
      by `POST /rooms`, `POST /join` and `POST /rooms/{id}/owner-session` and sent as
      `Authorization: Bearer <jwt>`. A JWT for room A is rejected on every route of room B.
    - Membership is re-checked on every request: kicked or banned members get `403` even with a
      JWT that has not expired yet.

    **Realtime** traffic goes over the WebSocket at `GET /ws`; see that operation and the
    `ClientFrame` / `ServerFrame` schemas.

    **Rate limits**: `POST /join` 10/min per IP; sending 5 messages/s per member;
    image uploads 10/min per member (bursts of 5). Exceeding one returns `429 rate_limited`.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8000/api
    description: Docker Compose (through Nginx, which strips the /api prefix)
  - url: http://localhost:8080
...
security:
  - roomJWT: []
paths:
...
  /rooms/{id}/invites:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    post:
      tags: [invites]
      operationId: createInvite
      summary: Create an invite link (host only)
      description: |
        Issues a random 128-bit token. Only its SHA-256 hash is stored, so the token and link are
        returned **once**. Share `invite_url` as a link or render it as a QR code. When the server
        runs without `APP_BASE_URL`, `invite_url` is a path (`/j/{token}`) and the client prefixes
        its own origin.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteRequest"
            example:
              expires_in_minutes: 1440
              max_uses: 20
      responses:
        "201":
          description: Invite created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedInvite"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenOrRemoved"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/RoomClosed"
    get:
      tags: [invites]
      operationId: listInvites
      summary: List usable invites (host only)
      description: Invites that are not revoked, expired or used up, newest first. Tokens are never returned again.
      responses:
        "200":
          description: Usable invites.
          content:
            application/json:
              schema:
                type: object
                required: [invites]
                properties:
                  invites:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invite"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenOrRemoved"
        "404":
          $ref: "#/components/responses/NotFound"
  /rooms/{id}/invites/{inviteId}:
    parameters:
      - $ref: "#/components/parameters/RoomId"
      - name: inviteId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      tags: [invites]
      operationId: revokeInvite
      summary: Revoke an invite (host only)
      description: The link stops working immediately. Members who already joined stay in the room. Revoking twice is a no-op.
      responses:
        "204":
          description: Invite revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenOrRemoved"
        "404":
          $ref: "#/components/responses/NotFound"
  /join:
    post:
      tags: [invites]
      operationId: joinRoom
      summary: Join a room with an invite token
      description: |
        Redeems an invite (atomically counted against `max_uses`) and creates a new member with the
        given nickname. Duplicate nicknames are allowed; clients label them by join order
        ("สมชาย", "สมชาย (2)"). Limited to 10 requests per minute per IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRequest"
            example:
              token: 5_m8xks2Bn7OKMiuCMO4pQ
              display_name: สมชาย
      responses:
        "200":
          description: Joined; the session is bound to the invite's room.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinedRoom"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "404":
          description: "`invite_invalid`: unknown or revoked token."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: { code: invite_invalid, message: invite is invalid or was revoked }
        "409":
          description: "`room_full`: the room already has 200 active members."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: "`invite_expired`, `invite_exhausted` or `room_closed`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: { code: invite_expired, message: invite has expired }
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /rooms/{id}/messages:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    get:
      tags: [messages]
      operationId: listMessages
      summary: Load message history
      description: |
        Returns one page, oldest first. Pass `next_cursor` as `before` to load the page before it;
        `next_cursor` is `null` when there is nothing older. Message ids are ULIDs and sort by time.
      parameters:
        - name: before
          in: query
          required: false
          description: Only messages older than this message id.
          schema:
            $ref: "#/components/schemas/MessageId"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: A page of messages.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePage"
        "400":
          $ref: "#/components/responses/InvalidInput"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Removed"
        "404":
          $ref: "#/components/responses/NotFound"
  /rooms/summaries:
    post:
      tags: [messages]
      operationId: summarizeRooms
      summary: Summaries for the room list
      description: |
        Takes every room session a browser holds and answers for each one independently, so one
        expired or revoked session does not fail the whole list. Each entry is authorised by its own
        room JWT in the body, so this route needs no Authorization header. At most 100 rooms.
        `unread` counts messages from **other** members after `last_read_id`, capped at 99.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SummariesRequest"
      responses:
        "200":
          description: One summary per requested room, in request order.
          content:
            application/json:
              schema:
                type: object
                required: [rooms]
                properties:
                  rooms:
                    type: array
                    items:
                      $ref: "#/components/schemas/RoomSummary"
        "400":
          $ref: "#/components/responses/InvalidInput"
...
  /ws:
    get:
      tags: [realtime]
      operationId: openWebSocket
      summary: Open the realtime WebSocket
      description: |
        Upgrade to WebSocket. Browsers cannot set headers on a WebSocket, so the JWT is sent as the
        **second subprotocol**, which keeps it out of URLs and access logs:

        ```js
        new WebSocket(`wss://host/ws?room=${roomId}`, ["chat", jwt])
        ```

        The server selects subprotocol `chat`. `Origin` must match the host (or `ALLOWED_ORIGINS`),
        otherwise the upgrade is refused with `403`. Authentication happens after the upgrade, so
        auth failures arrive as an `error` frame followed by a close code.

        **Flow**
        1. After `open`, send `resume` with the newest message id you have; missed messages are
           replayed as `message` frames, then live traffic continues. Without an id, reload the
           latest page over REST instead.
        2. Send `ping` every 25 s; a connection silent for 60 s is dropped.
        3. `send` is stored before it is broadcast. The sender gets `ack`; everyone in the room,
           the sender included, gets `message`. Re-sending the same `client_msg_id` (for example
           after a reconnect) returns the stored message and never duplicates it.
        4. To send an image, upload it first (`POST /rooms/{id}/images`) and put its id in
           `image_id`; `body` is then an optional caption.

        **Close codes**: `4001` unauthorized, `4003` kicked or banned, `4004` room closed; do not
        reconnect after these. Any other close: reconnect with exponential backoff (1, 2, 4, 8 … 30 s)
        and `resume`.

        **Frames** are JSON text; see `ClientFrame` and `ServerFrame`. Frames are limited to 16 KB.
      security: []
      parameters:
        - name: room
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: Sec-WebSocket-Protocol
          in: header
          required: true
          description: "`chat, <jwt>`"
          schema:
            type: string
      responses:
        "101":
          description: Switching Protocols; then exchange `ClientFrame` / `ServerFrame` messages.
        "403":
          description: Origin not allowed.
        "426":
          description: Not a WebSocket upgrade request.
components:
  securitySchemes:
    roomJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Room-bound JWT from `POST /rooms`, `POST /join` or `POST /rooms/{id}/owner-session`. Valid for 12 hours.
  parameters:
    RoomId:
      name: id
      in: path
      required: true
      description: Room id. Must match the room of the JWT.
      schema:
        type: string
        format: uuid
  responses:
    InvalidInput:
      description: "`invalid_input` (or `message_too_long`): the request is malformed; `message` explains why."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: { code: invalid_input, message: display_name is required }
    Unauthorized:
      description: "`unauthorized`: missing, expired or invalid JWT, or a JWT for another room."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: { code: unauthorized, message: missing or invalid credentials }
    Removed:
      description: "`kicked` or `banned`: the member was removed from the room."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: { code: banned, message: you are banned from this room }
    ForbiddenOrRemoved:
      description: "`forbidden` (not the host), `kicked` or `banned`."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: { code: forbidden, message: not allowed }
    NotFound:
      description: "`not_found`: the room, member, invite or image does not exist."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: { code: not_found, message: not found }
    RoomClosed:
      description: "`room_closed`."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: { code: room_closed, message: room is closed }
    RateLimited:
      description: "`rate_limited`."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: { code: rate_limited, message: too many requests }
  schemas:
    ErrorCode:
      type: string
      enum:
        - not_found
        - unauthorized
        - forbidden
        - invalid_input
        - invite_invalid
        - invite_expired
        - invite_exhausted
        - room_closed
        - room_full
        - rate_limited
        - banned
        - kicked
        - message_too_long
        - image_too_large
        - image_unsupported
        - internal
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: "#/components/schemas/ErrorCode"
            message:
              type: string
              description: Human-readable English detail; not for display logic.
    MessageId:
      type: string
      description: ULID; lexical order is time order.
      pattern: "^[0-9A-HJKMNP-TV-Z]{26}$"
      examples: [01J8Z3Q4R5S6T7V8W9XYZABCDE]
    DisplayName:
      type: string
      minLength: 1
      maxLength: 32
      description: Trimmed; inner whitespace collapsed. Counted in characters, so Thai text is fine.
    Role:
      type: string
      enum: [owner, member]
...
    Session:
      type: object
      required: [room_id, member_id, role, jwt, expires_at]
      properties:
        room_id:
          type: string
          format: uuid
        member_id:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/Role"
        jwt:
          type: string
        expires_at:
          type: string
          format: date-time
...
    JoinRequest:
      type: object
      required: [token, display_name]
      properties:
        token:
          type: string
          description: The `{token}` part of the invite link `/j/{token}`.
        display_name:
          $ref: "#/components/schemas/DisplayName"
    JoinedRoom:
      allOf:
        - $ref: "#/components/schemas/Session"
        - type: object
          required: [room_name]
          properties:
            room_name:
              type: string
...
    CreateInviteRequest:
      type: object
      properties:
        expires_in_minutes:
          type: integer
          minimum: 1
          maximum: 43200
          default: 1440
          description: 0 or omitted means the default of 24 hours.
        max_uses:
          type: [integer, "null"]
          minimum: 1
          maximum: 200
          description: null or omitted means unlimited.
    CreatedInvite:
      type: object
      required: [id, token, invite_url, expires_at, max_uses]
      properties:
        id:
          type: string
          format: uuid
        token:
          type: string
          description: Random 128-bit, base64url. Returned only here.
        invite_url:
          type: string
          description: Absolute when the server has `APP_BASE_URL`, otherwise the path `/j/{token}`.
          examples: [/j/5_m8xks2Bn7OKMiuCMO4pQ]
        expires_at:
          type: string
          format: date-time
        max_uses:
          type: [integer, "null"]
    Invite:
      type: object
      required: [id, expires_at, max_uses, used_count, created_at]
      properties:
        id:
          type: string
          format: uuid
        expires_at:
          type: string
          format: date-time
        max_uses:
          type: [integer, "null"]
        used_count:
          type: integer
        created_at:
          type: string
          format: date-time
...
    ImageRef:
      type: object
      required: [id, content_type, width, height]
      properties:
        id:
          type: string
          format: uuid
        content_type:
          type: string
          enum: [image/jpeg, image/png, image/gif]
        width:
          type: integer
        height:
          type: integer
...
    Message:
      type: object
      required: [id, member_id, display_name, client_msg_id, body, image, created_at]
      properties:
        id:
          $ref: "#/components/schemas/MessageId"
        member_id:
          type: string
          format: uuid
        display_name:
          type: string
          description: Name at send time; clients may relabel duplicates using the member list.
        client_msg_id:
          type: string
          format: uuid
        body:
          type: string
          maxLength: 2000
          description: May be empty when `image` is set. Plain text; render escaped, never as HTML.
        image:
          oneOf:
            - $ref: "#/components/schemas/ImageRef"
            - type: "null"
        created_at:
          type: string
          format: date-time
    MessagePage:
      type: object
      required: [messages, next_cursor]
      properties:
        messages:
          type: array
          description: Oldest first.
          items:
            $ref: "#/components/schemas/Message"
        next_cursor:
          oneOf:
            - $ref: "#/components/schemas/MessageId"
            - type: "null"
    SummariesRequest:
      type: object
      required: [rooms]
      properties:
        rooms:
          type: array
          maxItems: 100
          items:
            type: object
            required: [room_id, jwt]
            properties:
              room_id:
                type: string
                format: uuid
              jwt:
                type: string
              last_read_id:
                description: Newest message this client has seen; empty counts everything as unread.
                oneOf:
                  - $ref: "#/components/schemas/MessageId"
                  - type: string
                    maxLength: 0
    RoomSummary:
      type: object
      required: [room_id, state, last_message, unread]
      properties:
        room_id:
          type: string
          format: uuid
        state:
          type: string
          enum: [active, kicked, banned, unauthorized, not_found]
          description: Anything but `active` means this session can no longer use the room.
        name:
          type: string
          description: Present when `state` is `active`.
        status:
          type: string
          enum: [active, closed]
          description: Present when `state` is `active`.
        last_message:
          oneOf:
            - $ref: "#/components/schemas/Message"
            - type: "null"
        unread:
          type: integer
          minimum: 0
          maximum: 99

    # ---- WebSocket frames ----
...